% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cancer.ds}
\alias{cancer.ds}
\title{A demo microarray dataset from a cancer study}
\format{
A SummarizedExperiment object.
}
\usage{
data(cancer.ds)
}
\description{
A demo microarray dataset from a cancer study
}
\details{
This dataset is constructed using the first three columns of data **gse16873.d** in package **pathview** (i.e. columns "DCIS_1","DCIS_2" and "DCIS_3"). the original values were used without additional processing. It is constructed for showing SBGNview's visualization ability, not for data analysis. Each column in the assay table is a pair of cancer-v.s.-control samples. The value of a gene in a column is the log fold change of this gene in the corresponding pair of cancer-v.s.-control samples.
}
\keyword{datasets}
