% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IFNg}
\alias{IFNg}
\title{RNA-Seq result from a mouse IFNG knockout experiment}
\format{
A SummarizedExperiment object.
}
\usage{
data(IFNg)
}
\description{
RNA-Seq result from a mouse IFNG knockout experiment
}
\details{
This RNA-Seq dataset contains RNA abundance table of two groups: IFNG knockout mice and wild type mice. RNA abundance values were log2 transformed. For demo purpose, data of 4 IFNG knockout mice and 4 wild type mice were included.
The experiment and data processing was described in this work: Greer, Renee L., Xiaoxi Dong, et al. "Akkermansia muciniphila mediates negative effects of IFNG on glucose metabolism." Nature communications 7 (2016): 13329.
}
\keyword{datasets}
