% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapping}
\alias{mapping}
\alias{ENZYME_pathway.id}
\alias{hsa_KO_ENTREZID}
\alias{hsa_pathwayCommons_ENTREZID}
\alias{mmu_KO_ENSEMBL}
\alias{chebi_pathway.id}
\alias{mmu_KO_ENTREZID}
\alias{chebi_CompoundName}
\alias{CompoundName_pathwayCommons}
\alias{kegg.ligand_pathwayCommons}
\alias{hsa_pathwayCommons_ENSEMBL}
\alias{mmu_pathwayCommons_ENTREZID}
\alias{KO_pathway.id}
\alias{KO_pathwayCommons}
\alias{SYMBOL_pathway.id}
\alias{pathwayCommons_SYMBOL}
\alias{chebi_pathwayCommons}
\alias{mmu_pathwayCommons_ENSEMBL}
\alias{hsa_ENTREZID_pathwayCommons}
\title{Mapping table between two types of IDs}
\format{A matrix with two columns: the ID mapping between two types of IDs.}
\usage{
ENZYME_pathway.id

hsa_KO_ENTREZID

hsa_pathwayCommons_ENTREZID

mmu_KO_ENSEMBL

chebi_pathway.id

mmu_KO_ENTREZID

chebi_CompoundName

CompoundName_pathwayCommons

kegg.ligand_pathwayCommons

hsa_pathwayCommons_ENSEMBL

mmu_pathwayCommons_ENTREZID

KO_pathway.id

KO_pathwayCommons

SYMBOL_pathway.id

pathwayCommons_SYMBOL

chebi_pathwayCommons

mmu_pathwayCommons_ENSEMBL

hsa_ENTREZID_pathwayCommons
}
\description{
Mapping table between two types of IDs
}
\details{
Each dataset contains a mapping table. There are several types of ID pairs, such as molecule ID <=> pathway_glyph_ID, molecule ID <=> pathway ID, and molecule ID <=> KEGG ortholog ID. molecule ID <=> pathway_glyph_ID tables are extracted from Biopax files. For example: http://www.pathwaycommons.org/archives/PC3/v10/PathwayCommons10.reactome.BIOPAX.owl.gz.  Glyph IDs are extracted from the ID of each XML element "Protein". Its matching molecule ID is extracted from the corresponding XML child element "UnificationXref". See more details and examples in vignette 'SBGNview.data.vignette'
}
\keyword{datasets}
