% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathway.completeness.cutoff.info}
\alias{pathway.completeness.cutoff.info}
\title{Cutoffs of pathway completeness used for defining existance of pathway in a species}
\format{A matrix}
\usage{
pathway.completeness.cutoff.info
}
\description{
Cutoffs of pathway completeness used for defining existance of pathway in a species
}
\keyword{datasets}
