% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathway.species.pct_Mapped}
\alias{pathway.species.pct_Mapped}
\title{Pathway completeness of pathway in a species}
\format{A matrix}
\usage{
pathway.species.pct_Mapped
}
\description{
Pathway completeness of pathway in a species
}
\keyword{datasets}
