% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapping}
\alias{mapping}
\alias{ENZYME_pathway.id}
\alias{hsa_KO_ENTREZID}
\alias{mmu_KO_ENSEMBL}
\alias{chebi_pathway.id}
\alias{mmu_KO_ENTREZID}
\alias{chebi_compound.name}
\alias{compound.name_pathwayCommons}
\alias{kegg_pathwayCommons}
\alias{hsa_pathwayCommons_ENSEMBL}
\alias{mmu_pathwayCommons_ENTREZID}
\alias{KO_pathway.id}
\alias{KO_pathwayCommons}
\alias{SYMBOL_pathway.id}
\alias{pathwayCommons_SYMBOL}
\alias{chebi_pathwayCommons}
\alias{mmu_pathwayCommons_ENSEMBL}
\alias{hsa_ENTREZID_pathwayCommons}
\alias{chebi_kegg}
\alias{chebi_metacyc.SBGN}
\alias{compound.name_pathway.id}
\alias{compound.name_kegg}
\title{Mapping table between two types of IDs}
\format{
A matrix/data.frame with two columns: the ID mapping between two types of IDs.

An object of class \code{data.frame} with 16601 rows and 2 columns.

An object of class \code{data.frame} with 10446 rows and 3 columns.

An object of class \code{matrix} (inherits from \code{array}) with 11016 rows and 2 columns.

An object of class \code{data.frame} with 79911 rows and 2 columns.

An object of class \code{data.frame} with 11105 rows and 3 columns.

An object of class \code{data.frame} with 285207 rows and 2 columns.

An object of class \code{data.frame} with 298727 rows and 2 columns.

An object of class \code{data.frame} with 6448 rows and 2 columns.

An object of class \code{matrix} (inherits from \code{array}) with 383983 rows and 2 columns.

An object of class \code{matrix} (inherits from \code{array}) with 628413 rows and 2 columns.

An object of class \code{data.frame} with 47719 rows and 2 columns.

An object of class \code{data.frame} with 88148 rows and 2 columns.

An object of class \code{data.frame} with 252730 rows and 2 columns.

An object of class \code{data.frame} with 406940 rows and 3 columns.

An object of class \code{matrix} (inherits from \code{array}) with 60425 rows and 2 columns.

An object of class \code{matrix} (inherits from \code{array}) with 619656 rows and 2 columns.

An object of class \code{matrix} (inherits from \code{array}) with 384671 rows and 2 columns.

An object of class \code{data.frame} with 20692 rows and 2 columns.

An object of class \code{matrix} (inherits from \code{array}) with 60659 rows and 2 columns.

An object of class \code{data.frame} with 571283 rows and 2 columns.

An object of class \code{data.frame} with 61143 rows and 2 columns.
}
\usage{
data(ENZYME_pathway.id)

data(hsa_KO_ENTREZID)

data(mmu_KO_ENSEMBL)

data(chebi_pathway.id)

data(mmu_KO_ENTREZID)

data(chebi_compound.name)

data(compound.name_pathwayCommons)

data(kegg_pathwayCommons)

data(hsa_pathwayCommons_ENSEMBL)

data(mmu_pathwayCommons_ENTREZID)

data(KO_pathway.id)

data(KO_pathwayCommons)

data(SYMBOL_pathway.id)

data(pathwayCommons_SYMBOL)

data(chebi_pathwayCommons)

data(mmu_pathwayCommons_ENSEMBL)

data(hsa_ENTREZID_pathwayCommons)

data(chebi_kegg)

data(chebi_metacyc.SBGN)

data(compound.name_pathway.id)

data(compound.name_kegg)
}
\description{
Mapping table between two types of IDs
}
\details{
Each dataset contains a mapping table. There are several types of ID pairs, such as molecule ID <=> pathway_glyph_ID, molecule ID <=> pathway ID, and molecule ID <=> KEGG ortholog ID. molecule ID <=> pathway_glyph_ID tables are extracted from Biopax files. For example: http://www.pathwaycommons.org/archives/PC3/v10/PathwayCommons10.reactome.BIOPAX.owl.gz.  Glyph IDs are extracted from the ID of each XML element "Protein". Its matching molecule ID is extracted from the corresponding XML child element "UnificationXref". See more details and examples in vignette 'SBGNview.data.vignette'
}
\keyword{datasets}
