% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathway.species.pct_Mapped}
\alias{pathway.species.pct_Mapped}
\title{Pathway completeness in a species}
\format{
A matrix
}
\usage{
data(pathway.species.pct_Mapped)
}
\description{
Pathway completeness in a species
}
\details{
PathwayCommons only annotated human pathways, we mapped pathwayCommons' genes to other species using KEGG ortholog annotation. As a result, not all of the genes have corresponding genes in another species. We call the percentage of mapped genes the "coverage or completeness" in the species.
}
\keyword{datasets}
