% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sbgn.xmls}
\alias{sbgn.xmls}
\title{XML code of a SBGN-ML file}
\format{
A list of character strings
}
\usage{
data(sbgn.xmls)
}
\description{
XML code of a SBGN-ML file
}
\details{
Each string is the full XML code of a SBGN-ML file. It includes glyphs and arcs of a SBGN map. \strong{***Note:} Please note that \code{sbgn.xmls} is a large R object and will take a few seconds to load. It is necessary to be loaded into the environment in order to access the pre-generated SBGN-ML files.
}
\keyword{datasets}
